refactor: use `min` and `max` for finding elements
Several places in this file there were places where we'd loop over a list looking for the biggest or smallest value of something. This is what `max` and `min` do for us, and they handle all the bookkeeping so we don't have to.

This also adds type annotations to most of the file, but doesn't change anything about the project setup to enforce them yet.