build: don't require enumerating types packages
This makes webdriver's types available without having to explicitly enumerate it in `tsconfig.json`. Doing so seems to require us to enumerate some (but not all?) of the types packages we use. This way TS will see the webdriverio.d.ts file which then pulls in types from `@wdio/sync`.