feat: add endpoint to GET or POST audit board data
This endpoint allows updating the name and members of an audit board. For now, there's very little validation of the data. The database has a fixed number of audit board members, but this API endpoint will only return as many as are set.

When POSTing, all attributes are optional. If you want to set `members` you have to set exactly two for now.