feat: make `ballot-list` endpoint easier to use
Since the client needs to aggregate based on `audit_board`, this commit changes the `ballot-list` endpoint to accept an `audit_board_id` instead of a `batch_id`. The return structure changes to remove the `auditBoardId` and add `batch.{id,name,tabulator}`.

The update endpoint is unchanged, so the `batch.id` field should be used to make requests to update `vote` and `comment` on a `sampled_ballot`.