feat: add "Votes" field for contests
This is not really a complete feature in and of itself. This commit allows targeted contests to have a number of votes allowed but does not do anything with that number. It is included alongside "Number of Winners" in the audit report, but none of the actual audit math changes based on it.

The changes to actually use this data will have to come from someone else, likely @umbernhard.

Closes #189