fix(security): compare passwords in constant-time
Without this, the `/admin` route is vulnerable to timing attacks: https://en.wikipedia.org/wiki/Timing_attack

I realize this admin setup is intended to be short-lived, but I do not want to leave an obvious security vulnerability in the codebase for any length of time.

I could have used `safe_str_cmp` from `werkzeug.security`, but it just delegates to `hmac.compare_digest` when it's available. Since it was added in Python 3.3, we're fine to use it directly.