build(docker): fixes `docker-compose up`
The `db` container failed to start because the `9.6` image changed (presumably to the most recent `9.6.x` version) and now wants a password. Since we're only using this for development, I added the environment variable that makes postgres trust incoming requests. This is not suitable for a production environment, so I added a comment to make it clear what this file is for.