Add STATIC_FOLDER config variable to ensure tests don't rely on build
Previously, the tests relied on `arlo-client/build/` existing in order to
make sure the server was working. This directory doesn't exist until you
do a production build for the first time.

This change uses the source folder for static files in test/development
and only uses the built folder for production.