fix: make better use of `try`/`except`
A bare `except` is considered bad python form, as it probably catches more than intended. This narrows the `except` we're keeping to the specific error we're looking for.

This also removes some `try`/`except` statements that were unnecessary. According to the Flask docs, if `get_json` is unable to parse the JSON value from the request, it'll raise `BadRequest` just as we were doing. So we can just let it follow the default behavior.