Fix ballot-list endpoints to filter by round
In both ballot-list endpoints (per audit board and per jurisdiction),
the round_id parameter was being ignored, so all ballots for all rounds
of the audits were being returned. This adds a filter to the database
query to restrict results to the specified round.

This commit adds a new test for each endpoint with two audit rounds to
ensure this behavior.

It also updates some existing tests that tested the ordering of the
ballot lists. These tests tried to test across multiple rounds of the
audit when there was only one round, so here they are simplified to be
clearer.