fix: add authorization to a few endpoints
This restricts creating elections within an organization to audit admins in that organization. It also restricts the jurisdiction file upload API to audit admins for the election.

This also changes the data model to _not_ require an organization for each election. Any elections without an organization are accessible as they were before: by knowing the URL.