Change TargetedContest model to Contest, adding is_targeted flag
To support both targeted and opportunistic contests, the first step is
to change the TargetedContest data type to be generic to all contests.
We add `is_targeted` as a boolean flag to specify which contests are
targeted and which are opportunistic. There are no other types of
contests, so this should be sufficient.