Add GET /election/<election_id>/sample-sizes
This endpoint returns the estimated sample size options for the current
round (the frontend will only use this for the first round for now, but
we'll need the logic for later rounds as well, so it's implemented).
This is based on the `compute_sample_sizes` function from the
single-jurisdiction flow.