Add @with_jurisdiction_access route decorator
Adds a route decorator `@with_jurisdiction_access`. This decorator
requires jurisdiction admin (JA) authorization, checks that it matches
the given election/jurisdiction ids, and loads the election and
jurisdiction objects.

Also modifies `@with_election_access` to not take a user type and only
accept audit admin (AA) authorization.

To simplify permissions for the time being, AAs and JAs will not share
any routes. JA routes will always have a `jurisdiction_id` in the path
so we can easily check that they are trying to access the jurisdiction
for which they are authorized.