Add assert_ok test helper and typecheck tests
Adds an `assert_ok` test helper to check for successful API responses.
In order to get the assertion error messages to display well, we have to
do two things:
- Add `__tracebackhide__ = True` to the helper function, which treats it
as an inline assertion in the test error reporting
- Call `pytest.register_assert_rewrite("tests.helpers")`, which passes
through the assert error messages to the test error reporting

The latter change required making creating a `tests/__init__.py` file,
which had the bonus side-effect of causing `mypy` to actually run on the
tests for the first time (who knew?) So we also fix some type errors.