Add endpoint for jurisdiction admins to create audit boards
Adds an endpoint
`POST
/election/<election_id>/jurisdiction/<jurisdiction_id>/round/<round_id>/audit-board`
that jurisdiction admins can use to create audit boards.

When the audit boards are created, the sampled ballots for that round
are divvyed up as fairly as possible between the audit boards.

Also adds a db constraint that audit board names must be unique within a
jurisdiction for each round.