Add list rounds endpoint for jurisdiction admins
Adds `GET /election/<election_id>/jurisdiction/<jurisdiction_id>/round`
to list audit rounds when logged in as a jurisdiction admin.

This is pretty much the only piece of functionality that is identical
for audit admins and jurisdiction admins, but instead of reusing the
same endpoint, we make a new one. This has two benefits:
- All jurisdiction admin endpoints follow the same url scheme, which
includes the jurisdiction id. So we can always check that the user is
requesting info about district they are authorized for.
- Permissions are simpler to think about and implement. Every endpoint
can only be accessed by one user type.