Add AUDIT_BOARD user type and with_audit_board_access auth decorator
Adds a user type for audit boards. This user type is a bit different for
a couple of reasons:
- It doesn't correspond to one real world person, it corresponds to the entire
audit board.
- It's not represented by a User record in the database, but rather by
the AuditBoard record
- The audit board logs in by visting `/auditboard/<passphrase>`, instead
of using auth0 with email/password.

This change adds the new user type, implements the login flow for audit
boards, adds a `with_audit_board_access` decorator for routes that
should be restricted to audit board users, and updates the session
format and auth helpers to accommodate the differences outlined above.