Add ProxyFix middleware to set app http scheme to https
When our app is deployed on Heroku, the Heroku router terminates https
and sends requests to our app using http. This meant that our app
thought its url was using the http scheme rather than https, so when it
constructed redirect_uris for auth0, they used http.

Luckily, Heroku sets the `X-Forwarded-Proto` header to `https`, and this
middleware reads that header and configures the Flask app accordingly.

https://flask.palletsprojects.com/en/1.1.x/deploying/wsgi-standalone/?highlight=proxyfix#proxy-setups