Require jurisdiction admin authorization to upload ballot manifest
Ok, this commit is a bit complicated. The goal is to require
jurisdiction admin authorization for uploading/deleting ballot manifests
for the multi-jurisdiction flow, while leaving the existing endpoint for
the single-jurisdiction flow without authorization.

Here's how it's done:
- Factor out the logic from the existing endpoint `PUT/DELETE /election/<election_id>/jurisdiction/<jurisdiction_id>/manifest`
- Adding two corresponding endpoints, `PUT` and `DELETE
/election/<election_id>/jurisdiction/<jurisdiction_id>/ballot-manifest`
(yes, it's a very similar name - eventually I'm gonna come back and
namespace the endpoints for the two flows)
- Add another new endpoint to `GET` the results of the upload
- Reorganize some code - everything for single-jurisdiction goes into
`routes.py` (e.g. the `sample_ballots` helper) and everything for
multi-jurisdiction goes into a new file `ballot_manifest.py`, following
the established pattern.
- Update all of the tests for other multi-jurisdiction flow endpoints
that were relying on being able to upload the manifest without
authorization to use the new endpoint
- Add some new tests for the new endpoints