Only run tsc pre-commit hook when ts/tsx files change
`tsc` is pretty slow. We shouldn't have to wait for it to typecheck the
whole client if we only make changes to the server.

Note: we have to use `bash -c tsc` instead of just `tsc` because
`lint-staged` passed in an array of filenames to `tsc` but `tsc` has to
run on the whole project, so we want to ignore the arguments.