docs: fe_sendauth issue can require IPv6 trust
When running `make resetdb` on a Chromebook Crostini virtual environment with
the default Debian 10.2 Linux, I ran into the `fe_sendauth: no password supplied`
error noted in the README. But changing to `trust` for just IPv4 didn't resolve the issue,
When I changed to `trust` for IPv6 also, the make succeeded.
So I guess connections are being made via IPv6 in some configurations.
Note also that this was with Postgresql version 11, default on Debian.

The troubleshooting tip now also describes one way to reload the postgres configs,
which is also necessary.