fix: detect CSV file encoding (#531)
* fix: detect CSV file encoding

This allows files that do not have UTF-8 encoding to be uploaded. The check for confidence greater than zero is probably not needed, but I figured it wouldn't hurt since in that case we probably want to just try UTF-8.

* perf: only detect encoding if utf8 fails

`chardet.detect` is a bit slow (78ms on a 12K CSV on my MBP), so only use it if decoding as utf8 fails. This is a little slower in the minority case, but faster in the majority case.