/sample-sizes always return round one sample size options (#579)
This endpoint is only used for the audit setup flow. Currently, if you
navigate back to the Review & Launch screen after starting an audit, the
sample sizes shown will reflect the current state of the audit. This fix
ensures that they will always show the initial sample size options that
the user chose from for round 1 when starting the audit.