Simplify pytest command using pytest.ini
Adds a pytest.ini config file so that we don't have to put all the
default settings in the Makefile command. This allows running `pipenv
run pytest` if you want to add more flags to the command (`make` doesn't
allow you to pass in more flags, so we had a hacky way to work around
it before).