Order ballot draws in tests (#626)
The ordering of the ballot draws in the test helpers that audit the
ballots to run a round will make the tests more deterministic (since
sample sizes for future rounds depend on the audit results from previous
rounds). Currently, the tests are deterministic from run to run, but are
not resilient to changes in the code (e.g. if the way we sample ballots
changes and results in a different internal database ordering). We have
to make changes to the ballot sampling code, and in my exploration so
far, it broke these tests even though I'm fairly sure the logic didn't
change. So this PR sets us up to rely on these tests when changing that
code.