Report: add column for each contest's ticket numbers
When there are multiple targeted contests, we draw a separate sample for
each contest. That means a ballot might get sampled for multiple
contests. In the report, we want to be able to see which contests each
ballot was sampled for, so we create a column of ticket numbers
corresponding to each contest (previously, there was just one column for
all ticket numbers).