Remove database.cfg files (#658)
* Remove database.cfg files

The database.cfg file scheme is confusing and overly complex for the
amount of configuring we need to do. This change simplifies to configure
the database url through the existing env var `DATABASE_URL`. Reasonable
defaults are set in the config module for `development` and `test`
environments.

This obviates the need for a `config` directory, so we simplify to
`config.py`.

* README: Add missing env vars