Reports: refactor to functional style (#664)
The current report code has a bunch of helper functions that write to a
report stream. If one section was skipped, it resulted in an extra empty
line in between the report sections.

To fix this, I refactored the helpers to return a list of rows to write
(instead of writing them directly), so that the report building parent
function can easily deal with filtering out empty sections and joining
them correctly with new lines.