Run black only in necessary directories
And note in the Makefile that it will rewrite
all the Python files under those directories.

This helps prevent the uninitiated from having other files that aren't
part of the project abruptly reformatted.

That seems enough to take out the Troubleshooting note that I previously
put in the README.md.