Parallel tests (#675)
* Parallelize tests

Uses pytest-xdist to run the server tests in parallel. We use one database throughout most of the tests, which also simulates the real-world a bit better and helps ensure we are scoping data correctly.

In working on this PR, I found a race condition in bgcompute - when computing sample sizes, we can accidentally sample ballots twice for the same contest (for round > 1). So I copied the locking pattern from process_file to lock the RoundContest object.

* Simplify client fixture