scripts.resetdb: stamp db with latest migration revision
Whenever we reset the database schema, we need to mark in the database
the version of the schema we're using in order for Alembic to be able to
tell where to start from when running migrations on the database. I'd
been doing this ad hoc whenever I encountered a db without a revision
stamp, but I realized that would be less likely to happen if we stamp
every time we reset (or create) a database.