Clear data instead of dropping tables when seeding data for Cypress (#700)
Previously, we used the existing `resetdb` script to prepare the db for
seed data in Cypress tests. The problem with this approach is that it
hangs when a Flask app is already running, since that app has an open db
session, so SQLAlchemy gets blocked trying to drop the tables.

Instead, we add a new script `cleardb` that iterates over the tables and
deletes all the data from them. This does not block if the Flask app is
already running.

With this block removed, we can move the seeding of data into the actual
Cypress test spec so that it runs every time we run the tests, instead
of trying to run it via bash scripts for test running (much simpler and
less error-prone).