Multiple user types per route (#742)
Reworks the Flask route decorators that control permissions to allow multiple user types per route. This will be useful since we want to start letting AAs access more of the JA routes.

Also updates a ton of tests, which were lazily using an election that wasn't tied to an organization to circumvent the need to have the correct AA user logged in.