Convert test constants to snapshots
Most of these constants depend on the behavior of the sampler, which
changes in response to any changes in input (e.g. changing the way we
key batches, which is coming up in a later PR). Using snapshots ensures
that we can verify when these constants change without having to
manually update them by hand.