csv_parse: support composite unique keys (#788)
For ballot comparison audits, we are changing the ballot manifest to
have composite unique keys on the Tabulator and Batch Name columns -
i.e. each row must be uniquely identified by Tabulator+Batch Name.

This PR modifies the csv_parse module to support this case. Since we
never actually want two columns to both be independently required to be
unique (and I can't think of a situation in which we will), we modify
the behavior to make a composite unique key out of any columns which are
specified as `unique=True`.

While I was at it, I also removed the constraint that columns must in a
particular order, since I figured people might add the
Tabulator/Container columns at the beginning or at the end. We only
required ordered columns because there were some assumptions that the
columns were ordered in the parser, so these changes should make the
parser more robust.