Tabulator/Container in manifest for ballot comparison audits (#789)
* Tabulator/Container in manifest for ballot comparison audits

Product changes:
- For ballot comparison audits, batches are not uniquely identified by
Batch Name alone, but rather by Tabulator+Batch Name (this is how they
are recorded in the CVR). So we require the Tabulator column in the
ballot manifest and use Tabulator+Batch Name as the unique identifier
for batches throughout the audit.
- For ballot comparison audits, multiple batches of ballots might be
stored in one physical container. We allow an optional Container column
in the ballot manifest. If Container is provided, when divvying up the
sampled ballots between audit boards, we ensure that all ballots in the
same Container are given to the same audit board.
- These changes may both be useful in other types of audits as well, so
we make this functionality available (though not required) for other
audit types.

Implementation notes:
- Removed the Batch.storage_location field (unused) and replaced it with
Batch.container
- If Tabulator or Container is specified, we require that it must be
specified for every batch (all rows in the manifest)
- Changed the batch key for sampling ballots to `(batch.tabulator,
batch.name)`. If `batch.tabulator` is `None`, then this is logically
equivalent to using `batch.name` as the batch key, so it should work
both when tabulator is or isn't given. However, this does change the
output of the sampler for existing test cases, since the sampler hashes
the given batch keys, so we do expect many of the test snapshots to
change, but only in number of samples, not in other logical results.

* Add Tabulator to audit report