Include ballots that are missing contest in sample_cvr
Currently, when constructing the `sample_cvr` object to pass to the
`supersimple` audit math, we omit any ballot that the audit board did
not record any interpretation for (i.e. the contest was missing from
that ballot).

This impacts `supersimple.compute_risk`, since it iterates over the
`sample_cvr` to compute the p-value.