Count not found ballots as two-vote overstatement
If an audit board can't find a ballot, the research says we should
conservatively count that as a two-vote overstatement.

We pass `None` as the `sample_cvr` value for this case, and then
interpret that in `compute_discrepancies`.

Note: the typechecker is not smart enough to deal with this change to
the type, so there are also some other changes to accommodate that.