Expire sessions (#864)
* Expire sessions

Configures Flask to expire session cookies after 2 hours. Also updates
the frontend to display a message when this happens (note: we just
display the message when the frontend gets a 401, which could also
happen from trying to access a route while not logged in for some other
reason, but this is the most common case so I figured why not).