Show Auth0 login errors to users
If something goes wrong in the OAuth flow, Auth0 sends an error message
to the callback. We catch those errors and redirect the user to the
login screen, displaying the error.

Two changes to make this possible:
- Standardize the way we show pass messages to the login screen
- Change the auth tests to use the `patch` method of the mocking module,
which automatically clears the mocks out once the test is done
(previously, they were carrying over between tests)