Sampler manifest list of ballot positions
Change the interface to the sampler to take a list of ballot positions
instead of just a number of ballots. This will allow us to use the CVR
to generate the actual list of ballot positions (which aren't always
consecutive) and prevent sampling ballots that don't have the given
contest on them.