Sampler manifest from CVR
In ballot comparison audits, we should only allow the sampler to choose
from ballots that had a result recorded in the CVR for the targeted
contest. So we load the CvrBallots for all the targeted contests
jurisdictions, and then parse out whether or not each ballot had a
result recorded for each contest when building the sampler manifest.