Move state input from Participants to Settings screen
There are some issues going on with the two file uploads in the
Participants screen in ballot comparison audits. This is the first step
to simplify that form so that we can fix the issues there by taking out
this extra field that never really belonged there anyway.

Changes:
- Move the state input
- Remove the localization scheme for states (there's no need for
localization - we're probably only going to use Arlo in the US in
English for now)
- Restyle the settings screen to be more consistent
- Add an integration test for the settings screen
- Add a percent sign to the risk limit display values