Fix bgcompute test race condition (#942)
* Fix bgcompute test race condition

Add an optional election_id filter to bgcompute functions so that they
can be isolated to the specific election being processed in each test
case. This will prevent race conditions when running the tests in
parallel (one test would run a bgcompute function that would prematurely
process a file from another test).

* Put election id in test juridsiction admin email

Add the election_id to the jurisdiction admin emails created for each test case to avoid clashes when creating the users (which now happens more often as a result of the change to bgcompute).