Don't default a sample size option as selected after audit launches
They might not have selected the first sample size option, so this is
misleading.

Ideally, we'd send back the option they selected when launching the
audit, but currently we're only storing the sample size number, not
which option they picked (and the options aren't guaranteed to have a
unique number), so this would be a bigger change.