Fix bug for standardized contests with a dot in their name
The standardized contests selection form was breaking when a contest had a dot in its name. This is because Formik uses dot notation field names to access fields, both in the Field `name` prop and in `setFieldValue`.

The recommended fix in the Formik docs is to use the bracket notation:
`['field']`, but this won't work if the field name has apostrophes in
it.

This fix removes uses of the `name` prop and replaces `setFieldValue`
with `setValues`.