Store which sample size option was selected
When launching the audit, we used to just store the sample size number that was selected. This meant that there was no way to display which sample size option corresponding with that number had been chosen (e.g. was it "asn" or was it "custom"?)

This changes the create round endpoint to take the entire sample size option and store it in the RoundContest object. It also makes the /sample-sizes endpoint return which sample sizes were selected for the first round so we can show them to the user.

This also opens up the ability for more complex sample size option formats, which we'll need for hybrid audits.