Set contest metadata from CVRs on write (#1069)
Previously, we computed contest metadata (e.g. total votes, votes for each
choice) based on the CVRs at read time: either when computing the sample size
options or finally when starting the round.

Two issues with this:
1. A bug! There was a guard to make sure we only computed the contest metadata once, which meant that if a jurisdiction changed their CVR after the first time sample sizes were computed, it wouldn't be taken into account.
2. The upcoming redesign to the Review screen shows this computed contest metadata via the /contests endpoint. That meant that both the sample sizes endpoint and the contests endpoint, which both get called by the review screen, were going to need to invoke this computation, so it seemed simpler to just move it to be computed on write so that we didn't have multiple different readers trying to lazily compute the same data. Especially since the contests endpoint is called in a variety of other circumstances besides the Review screen.

This change also factors out the validation of whether we are ready to compute the contest metadata (i.e. whether all the CVRs have been uploaded) so that the sample sizes endpoint can continue to raise that error.