Hybrid audit: sample size integration (#1076)
* Hybrid audit: sample size integration

Integrate with the sample size math for hybrid audits (a dummy version). There is one big design decision to think through here: We need to separate the audit results for CVR ballots and non-CVR ballots.

For CVR ballots, the math needs a CVR-like dictionary with the audit result for each sampled ballot. We already compute this for ballot comparison, so we simply add a filter to make sure its only looking at CVR ballots.

For non-CVR ballots, it's a bit trickier. For ballot polling, we compute the vote counts for each contest choice at the end of each round and store them in the RoundContestResult object. Here, I chose to add a filter during that process to only count non-CVR ballots, so the resulting counts only apply to that segment.

The downside to this is that we won't be storing a cumulative total across both segments. But that's not a big deal, since the math never actually requires the cumulative total. We may want to show it in the audit report, but it's easy to compute by adding together the totals from the two segments.

I do think this makes things a bit confusing, since certain database objects have different meanings in different audit types. I'm curious about exploring ideas to pay back some of this "design debt" that we have incurred by adding the audit methods one-by-one. One direction that might be fruitful is to explore where database objects are being reused in different ways across audit types and consider separating them out into separate tables specific to an audit type.