Factor out discrepancy calculation (#1096)
This enables the same discrepancy calculation to be used by ballot comparison and hybrid audits.

It also changes where we ignore understatements. Now, understatements are computed so they can be shown in the audit report, but they are skipped when computing sample sizes or p-values.

Also removes the unused `discrepancy_cvr` field from the Discrepancy dict.