Compute ballot position for CVR ballots based on record id (#1154)
The record ids in for each batch in the CVR might not always be a continuous sequence - there might be missing numbers. So if we assume that the record id is equivalent to the ballot's physical "position" in the batch, we might sample ballots that we can't find in the CVR.

This PR separates the concept of a record id and a ballot position. After loading the CvrBallots into the db, we order all the record ids for each batch and assign them a ballot position based on the index in that ordering.

Also adds a script to generate a manifest for a CVR.