Hide CVR/non-CVR vote counts when CVRs invalid
If there's a CVR validation error (which is returned by the sample size endpoint), the CVR/non-CVR vote totals returned by the contests endpoint will be null, which was causing an error. In that case, we just show empty cells in the vote counts table until the validation error is resolved.