Fix migration order
I messed up when rebasing the background sample sizes pull request and put its migration before the CVR ballot migrations in the migration ordering. That meant that it didn't get run automatically on deploy to staging. This should put the background sample sizes migration at the end of the ordering, so it should get run.