Make background sample sizes backwards compatible
Existing audits that were launched before we made a background task for samples sizes won't have a background task. We now expect that after the audit launches, we can load the previously computed sample size options from the cached value from the background task.

This PR changes the sample sizes endpoint to start a background taskafter the audit launches if there isn't one already in order to backfill this data for existing audits.