Change batch tallies validation errors to UserErrors (#1161)
When moving sample size option computation to a background task, I forgot to change the error type for batch tallies validation, since it was hidden in the batch_tallies function. I moved it out to a separate function and changed the error type so that it will be correctly stored and shown to the user (instead of crashing the background task).