Remove special case for round one sample sizes
Previously, we needed to be able to compute the round one sample sizes at any time in order to show the round one sample sizes on the Review screen even after the audit launched. Now, since we store the round one sample sizes after computing them in the background, we don't need this anymore.