Show all lines of imprinted ID in placeholders
Certain long imprinted ids were getting cut off. The splitTextToSize function was splitting on the space between "Imprinted ID:" and the long imprinted id, causing the id to be moved to the second line (which we then didn't print).

It's still not totally clear to me why splitTextToSize did that (especially because it did the right thing and cut off the imprinted id in the middle of the id for the labels on the same audit), but this is a viable fix since we have plenty of space on the placeholders.