Refactor CSVFile component  to prevent extra rerenders  (#1266)
* Refactor CSVForm component to prevent extra rerenders

Previously, the CSVForm component was set up to use a key to force the form to rerender every time we reloaded a new file from the server. This was proving to be problematic in some other changes to the form I was working on, so I figured out a way to preserve the existing behavior without rerendering the entire form on every server file reload.