Record user info in file upload activity
Since file uploads happen in the background worker, they don't have access to the session of the request that initiated the upload, so when recording the file upload activity, they couldn't report which user uploaded the file. To fix this, we pass the user info in as arugments to the background task.