Delete unused CvrBallots after audit completes
During a ballot comparison or hybrid audit, we need every record from the CVRs on hand in the database in case the corresponding ballot gets sampled. Once the audit completes, we only need the CVRs for the ballots that got sampled. In this change, at the end of the audit, we delete all CvrBallots that don't correspond to a sampled ballot. Since CvrBallots are the largest use of database storage by an order of magnitude, this will hopefully significantly slow database growth.