Bugfix: load audits separately from user data (#1287)
* Bugfix: load audits separately from user data

Previously, we loaded the list of an audit admin's audits with their user data. This caused complications in a few different places, which we previously hacked around. A bug appeared when leaving an audit admin page open when logged in as a different user type, and then trying to load the home screen for the audit admin page again.

To fix this, we create a separate endpoint for loading the list of audits and decouple that from loading the user data. This impacts:
- loading the list of audits on login
- reloading the list of audits after deleting an audit
- *not* reloading the list of audits after creating an audit