Check for DATABASE_URL in Heroku release phase command
When creating review apps, sometimes the database isn't provisioned until after the first build and release are executed, which causes a release phase command failure. This change adds a guard to the release phase command to prevent that case. It shouldn't impact production at all because the database will always be provisioned.