Recompute contest metadata when standardized contests change
If standardized contests change after contests are already created, the contest universe for those contests will change (which we already handled). If the contest universe changes, however, that means that any contest metadata that was computed from files uploaded in that universe (manifests, CVRs) also needs to be updated.

This change ensures that we recompute that metadata. It also factors out the recomputing of metadata functionality into one function to make sure that it happens consistently whenever any dependencies change.