Fix flaky file upload tests
These tests were failing every once in a while. I'm not 100% sure why, but I think it had something to do with polling for file processing status. So I just modified the tests so that they don't have any polling loops (by changing the API response to not have a "processing" phase and skip straight to "processed"). The test should still test the same thing, but hopefully not flake out.