Always use react-query mutateAsync to ensure form spinners work correctly
react-query mutations have two main methods: `mutate`, which takes an `onSuccess` callback, and `mutateAsync`, which returns a Promise. I had experimented with using both in different places where I thought appropriate, but I realized that in order to play nicely with react-hook-form, we need to always use `mutateAsync`. The reason why is that we have a pattern of using `formState.isSubmitting` to show a spinner on the submit button while the form submission is in progress. `formState.isSubmitting` is controlled by `handleSubmit(onSubmit)`, which waits for the provided `onSubmit` function to complete. When using `mutate`, `onSubmit` returns immediately, and thus `formState.isSubmitting` is immediately false, and we don't get any spinner.