Don't crash worker on caught task errors
Any error that occurs during a task that we can catch and save to the
database, we should send to Sentry but not crash the worker, since it's
safe to continue working. Any error that happens in the rest of the
worker outside of the task context can still crash as before.